heres the starting base code I wrote
Basically went through MSEBase code and only took the things we needed.
I added and tweeked variable names to suit us better.
this is incomplete as not all motors/encoders/sensors have been added
and the pin numbers might change. but we'll only have to change it once
at the top instead of going through the whole code.
#goTeam